Update nuget depependencies
Refactor/restructure
