+++ MVCForum.IOC/packages.config
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Common.Logging" version="3.0.0" targetFramework="net45" />
-  <package id="Common.Logging.Core" version="3.0.0" targetFramework="net45" />
+  <package id="Common.Logging" version="3.2.0" targetFramework="net45" />
+  <package id="Common.Logging.Core" version="3.2.0" targetFramework="net45" />
   <package id="CommonServiceLocator" version="1.3" targetFramework="net45" />
   <package id="EntityFramework" version="6.1.3" targetFramework="net45" />
   <package id="Microsoft.AspNet.Mvc" version="5.2.3" targetFramework="net45" />
