+++ MVCForum.Website/Areas/Admin/AdminAreaRegistration.cs
@@ -18,21 +18,21 @@ namespace MVCForum.Website.Areas.Admin
             RouteTable.Routes.LowercaseUrls = true;
             RouteTable.Routes.AppendTrailingSlash = true;
 
-            context.MapRoute(
-                "Admin_editcategoryroute",
-                "Admin/{controller}/{action}/{id}",
-                new { controller = "AdminCategory", action = "Index", id = UrlParameter.Optional }
-            );
-            context.MapRoute(
-                "Admin_edituserroute",
-                "Admin/{controller}/{action}/{userId}",
-                new { controller = "Admin", action = "Index", userId = UrlParameter.Optional }
-            );
-            context.MapRoute(
-                "Admin_pagingroute",
-                "Admin/{controller}/{action}/{page}",
-                new { controller = "Account", action = "Index", page = UrlParameter.Optional }
-            );
+            //context.MapRoute(
+            //    "Admin_editcategoryroute",
+            //    "Admin/{controller}/{action}/{id}",
+            //    new { controller = "AdminCategory", action = "Index", id = UrlParameter.Optional }
+            //);
+            //context.MapRoute(
+            //    "Admin_edituserroute",
+            //    "Admin/{controller}/{action}/{userId}",
+            //    new { controller = "Admin", action = "Index", userId = UrlParameter.Optional }
+            //);
+            //context.MapRoute(
+            //    "Admin_pagingroute",
+            //    "Admin/{controller}/{action}/{page}",
+            //    new { controller = "Account", action = "Index", page = UrlParameter.Optional }
+            //);
             context.MapRoute(
                 "Admin_defaultroute",
                 "Admin/{controller}/{action}/{id}",
