+++ MVCForum.Website/packages.config
@@ -2,8 +2,8 @@
 <packages>
   <package id="AjaxMin" version="4.48.4489.28432" targetFramework="net40" />
   <package id="Akismet.NET" version="1.0" targetFramework="net40" />
-  <package id="Common.Logging" version="3.0.0" targetFramework="net45" />
-  <package id="Common.Logging.Core" version="3.0.0" targetFramework="net45" />
+  <package id="Common.Logging" version="3.2.0" targetFramework="net45" />
+  <package id="Common.Logging.Core" version="3.2.0" targetFramework="net45" />
   <package id="CommonServiceLocator" version="1.3" targetFramework="net45" />
   <package id="DataAnnotationsExtensions" version="1.1.0.0" targetFramework="net40" />
   <package id="DataAnnotationsExtensions.MVC3" version="1.1.0.0" targetFramework="net40" />
